+++ arch/x86/drivers/keyboard.rs
@@ -47,7 +47,7 @@ fn keypress(code: u8) {
         (0x46, true) => led(0b001), // Scroll lock
         (c, true) if c < 0x3A => unsafe {
             // handle character
-            let mut ch = if shift { LayoutShift[c] } else { Layout[c] };
+            let mut ch = if shift { LayoutShift[c as uint] } else { Layout[c as uint] };
             if ch != 0 {
                 if caps_lock && isalpha(ch) {
                     ch ^= 1 << 5;
