+++ arch/x86/io/mod.rs
@@ -15,13 +15,13 @@ unsafe fn write_char(c: char) {
     if c == '\x08' {
         if pos > 0 {
             if pos % 80 == 0 {
-                while (*vga::SCREEN)[pos-1].char == 0 {
+                while (*vga::SCREEN)[(pos-1) as uint].char == 0 {
                     pos -= 1;
                 }
             }
             else if pos > 0 {
                 pos -= 1;
-                (*vga::SCREEN)[pos].char = 0;
+                (*vga::SCREEN)[pos as uint].char = 0;
             }
         }
     }
@@ -32,7 +32,7 @@ unsafe fn write_char(c: char) {
         seek(4 - pos % 4);
     }
     else {
-        (*vga::SCREEN)[pos].char = c as u8;
+        (*vga::SCREEN)[pos as uint].char = c as u8;
         pos += 1;
     }
